web-ui/seekBar: implement hybrid interval/observation updating.
The problem with simply observing the percent-pos property is that its
update frequency is dependent on the framerate. However, the problem
with using an interval is that running it when paused wastes cycles
unnecessarily. By combining the two, good behavior can be had in both
the pause and play states.