Add VpmStandardWindowButtonView class.
This class allows placing the standard traffic light window buttons in
any window, regardless of the existence of a titlebar. The initial plan
was to reproduce the style of these buttons with SVG in the WebView.
However, this idea proved to be infeasible due to the number of
permutations in appearance that the buttons can have.