web-ui: fix missing button icons.
At some point webkit flat out broke applying css filters to inline svgs.
For some weird reason, the filter on the button:hover div in main.styl
works a-ok, I'm guessing because it's applied to the div itself rather
than directly to the svg. Unfortunately, that filter clobbers this one,
which is why they were in different scopes.