web ui: rejigger hover time.
This fixes the problem where the hover time pop-in would render
underneath the progress bar if hit during the hover transition. Now it
is a child of the transitioning div so it's always at the right hight,
though it doesn't translate smoothly like everything else for some
reason. The judder isn't really noticeable but should be investigated
at some point.